Swift: Create a MX3PID struct to encapsulate address and medium data. Misc housekeeping.
